+++ src/pages/Tickets/styles.js
@@ -129,4 +129,6 @@ export const LowLevel = styled.div`
     }
 `
 
+
+
 export default Container;
